+++ components/molecule/buttonGroupField/demo/index.js
@@ -1,4 +1,4 @@
-import {H1} from '@s-ui/documentation-library'
+import {H1, Paragraph, Code, Anchor} from '@s-ui/documentation-library'
 
 import ArticleInformation from './ArticleInformation.js'
 import ArticleSemantic from './ArticleSemantic.js'
@@ -8,12 +8,20 @@ const BASE_CLASS_DEMO = 'DemoMoleculeButtonGroupField'
 const DemoMoleculeButtonGroupField = () => {
   return (
     <div className="sui-StudioPreview">
-      <div className="sui-StudioPreview-content sui-StudioDemo-preview">
-        <H1>Button Group Field</H1>
-        <ArticleInformation className={BASE_CLASS_DEMO} />
-        <br />
-        <ArticleSemantic className={BASE_CLASS_DEMO} />
-      </div>
+      <H1>Button Group Field</H1>
+      <Paragraph>
+        <Code>MoleculeButtonGroupField</Code> is the combination of and{' '}
+        <Anchor href="/workbench/molecule/buttonGroup">
+          <Code>MoleculeButtonGroup</Code>
+        </Anchor>{' '}
+        and a{' '}
+        <Anchor href="/workbench/molecule/field">
+          <Code>MoleculeField</Code>
+        </Anchor>
+      </Paragraph>
+      <ArticleInformation className={BASE_CLASS_DEMO} />
+      <br />
+      <ArticleSemantic className={BASE_CLASS_DEMO} />
     </div>
   )
 }