+++ components/molecule/dropdownOption/CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG
 
+# 1.42.0 (2022-03-11)
+
+
+### Bug Fixes
+
+* **components/molecule/dropdownOption:** fix default checkbox configuration ([32b55b2](https://github.com/SUI-Components/sui-components/commit/32b55b2daae1c76022b13cdfaa6b2cab2f2aa420))
+
+
+
 # 1.41.0 (2022-03-10)